Show error message when scheme is unsupported + updated readme
When the color scheme is not a raw file on the filesystem, show a message in the status area.

Also, created a method for getting the path of the current scheme and cleaned up some copy/pasted code.

Changed a bunch of instances of 'theme' to 'color scheme' for accuracy.